+++ example/lib/main-cupertino.dart
@@ -5,22 +5,24 @@
 import 'package:flutter/material.dart';
 import 'package:upgrader/upgrader.dart';
 
-void main() => runApp(MyApp());
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+
+  // Only call clearSavedSettings() during testing to reset internal values.
+  await Upgrader.clearSavedSettings(); // REMOVE this for release builds
+
+  // On Android, setup the Appcast.
+  // On iOS, the default behavior will be to use the App Store version of
+  // the app, so update the Bundle Identifier in example/ios/Runner with a
+  // valid identifier already in the App Store.
+  runApp(MyApp());
+}
 
 class MyApp extends StatelessWidget {
-  MyApp({
-    Key key,
-  }) : super(key: key);
+  MyApp({Key key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    // Only call clearSavedSettings() during testing to reset internal values.
-    Upgrader().clearSavedSettings(); // REMOVE this for release builds
-
-    // On Android, setup the Appcast.
-    // On iOS, the default behavior will be to use the App Store version of
-    // the app, so update the Bundle Identifier in example/ios/Runner with a
-    // valid identifier already in the App Store.
     final appcastURL =
         'https://raw.githubusercontent.com/larryaasen/upgrader/master/test/testappcast.xml';
     final cfg = AppcastConfiguration(url: appcastURL, supportedOS: ['android']);
@@ -28,13 +30,13 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Upgrader Example',
       home: Scaffold(
-          appBar: AppBar(
-            title: Text('Upgrader Example'),
-          ),
+          appBar: AppBar(title: Text('Upgrader Example')),
           body: UpgradeAlert(
-            appcastConfig: cfg,
-            debugLogging: true,
-            dialogStyle: UpgradeDialogStyle.cupertino,
+            Upgrader(
+              appcastConfig: cfg,
+              debugLogging: true,
+              dialogStyle: UpgradeDialogStyle.cupertino,
+            ),
             child: Center(child: Text('Checking...')),
           )),
     );