+++ test/upgrader_test.dart
@@ -27,8 +27,6 @@ void main() {
   const kEmptyPreferences = <String, dynamic>{};
 
   setUp(() async {
-    Upgrader.resetSingleton();
-
     // This idea to mock the shared preferences taken from:
     /// https://github.com/flutter/plugins/blob/master/packages/shared_preferences/test/shared_preferences_test.dart
     sharedPrefsChannel.setMockMethodCallHandler((MethodCall methodCall) async {
@@ -43,7 +41,7 @@ void main() {
       return null;
     });
     preferences = await SharedPreferences.getInstance();
-    await Upgrader().clearSavedSettings();
+    await Upgrader.clearSavedSettings();
   });
 
   tearDown(() async {
@@ -52,9 +50,7 @@ void main() {
 
   testWidgets('test Upgrader class', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader();
-    upgrader.platform = TargetPlatform.iOS;
-    upgrader.client = client;
+    final upgrader = Upgrader(platform: TargetPlatform.iOS, client: client);
 
     expect(tester.takeException(), null);
     await tester.pumpAndSettle();
@@ -95,10 +91,8 @@ void main() {
 
   testWidgets('test UpgradeWidget', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader();
-    upgrader.platform = TargetPlatform.iOS;
-    upgrader.client = client;
-    upgrader.debugLogging = true;
+    final upgrader = Upgrader(
+        platform: TargetPlatform.iOS, client: client, debugLogging: true);
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -128,17 +122,17 @@ void main() {
 
     expect(upgrader.messages, isNotNull);
 
-    expect(upgrader.messages!.buttonTitleIgnore, 'IGNORE');
-    expect(upgrader.messages!.buttonTitleLater, 'LATER');
-    expect(upgrader.messages!.buttonTitleUpdate, 'UPDATE NOW');
+    expect(upgrader.messages.buttonTitleIgnore, 'IGNORE');
+    expect(upgrader.messages.buttonTitleLater, 'LATER');
+    expect(upgrader.messages.buttonTitleUpdate, 'UPDATE NOW');
 
     upgrader.messages = MyUpgraderMessages();
 
-    expect(upgrader.messages!.buttonTitleIgnore, 'aaa');
-    expect(upgrader.messages!.buttonTitleLater, 'bbb');
-    expect(upgrader.messages!.buttonTitleUpdate, 'ccc');
+    expect(upgrader.messages.buttonTitleIgnore, 'aaa');
+    expect(upgrader.messages.buttonTitleLater, 'bbb');
+    expect(upgrader.messages.buttonTitleUpdate, 'ccc');
 
-    await tester.pumpWidget(const _MyWidget());
+    await tester.pumpWidget(_MyWidget(upgrader: upgrader));
 
     expect(find.text('Upgrader test'), findsOneWidget);
     expect(find.text('Upgrading'), findsOneWidget);
@@ -149,31 +143,29 @@ void main() {
 
     expect(upgrader.isTooSoon(), true);
 
-    expect(find.text(upgrader.messages!.title), findsOneWidget);
+    expect(find.text(upgrader.messages.title), findsOneWidget);
     expect(find.text(upgrader.message()), findsOneWidget);
     expect(find.text('Release Notes:'), findsOneWidget);
     expect(find.text(upgrader.releaseNotes!), findsOneWidget);
-    expect(find.text(upgrader.messages!.prompt), findsOneWidget);
+    expect(find.text(upgrader.messages.prompt), findsOneWidget);
     expect(find.byType(TextButton), findsNWidgets(3));
-    expect(find.text(upgrader.messages!.buttonTitleIgnore), findsOneWidget);
-    expect(find.text(upgrader.messages!.buttonTitleLater), findsOneWidget);
-    expect(find.text(upgrader.messages!.buttonTitleUpdate), findsOneWidget);
+    expect(find.text(upgrader.messages.buttonTitleIgnore), findsOneWidget);
+    expect(find.text(upgrader.messages.buttonTitleLater), findsOneWidget);
+    expect(find.text(upgrader.messages.buttonTitleUpdate), findsOneWidget);
 
-    await tester.tap(find.text(upgrader.messages!.buttonTitleUpdate));
+    await tester.tap(find.text(upgrader.messages.buttonTitleUpdate));
     await tester.pumpAndSettle();
-    expect(find.text(upgrader.messages!.buttonTitleIgnore), findsNothing);
-    expect(find.text(upgrader.messages!.buttonTitleLater), findsNothing);
-    expect(find.text(upgrader.messages!.buttonTitleUpdate), findsNothing);
+    expect(find.text(upgrader.messages.buttonTitleIgnore), findsNothing);
+    expect(find.text(upgrader.messages.buttonTitleLater), findsNothing);
+    expect(find.text(upgrader.messages.buttonTitleUpdate), findsNothing);
     expect(called, true);
     expect(notCalled, true);
   }, skip: false);
 
   testWidgets('test UpgradeWidget Cupertino', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader();
-    upgrader.platform = TargetPlatform.iOS;
-    upgrader.client = client;
-    upgrader.debugLogging = true;
+    final upgrader = Upgrader(
+        platform: TargetPlatform.iOS, client: client, debugLogging: true);
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -203,19 +195,18 @@ void main() {
 
     expect(upgrader.messages, isNotNull);
 
-    expect(upgrader.messages!.buttonTitleIgnore, 'IGNORE');
-    expect(upgrader.messages!.buttonTitleLater, 'LATER');
-    expect(upgrader.messages!.buttonTitleUpdate, 'UPDATE NOW');
+    expect(upgrader.messages.buttonTitleIgnore, 'IGNORE');
+    expect(upgrader.messages.buttonTitleLater, 'LATER');
+    expect(upgrader.messages.buttonTitleUpdate, 'UPDATE NOW');
 
     upgrader.messages = MyUpgraderMessages();
 
-    expect(upgrader.messages!.buttonTitleIgnore, 'aaa');
-    expect(upgrader.messages!.buttonTitleLater, 'bbb');
-    expect(upgrader.messages!.buttonTitleUpdate, 'ccc');
+    expect(upgrader.messages.buttonTitleIgnore, 'aaa');
+    expect(upgrader.messages.buttonTitleLater, 'bbb');
+    expect(upgrader.messages.buttonTitleUpdate, 'ccc');
+    upgrader.dialogStyle = UpgradeDialogStyle.cupertino;
 
-    await tester.pumpWidget(const _MyWidget(
-      dialogStyle: UpgradeDialogStyle.cupertino,
-    ));
+    await tester.pumpWidget(_MyWidget(upgrader: upgrader));
 
     expect(find.text('Upgrader test'), findsOneWidget);
     expect(find.text('Upgrading'), findsOneWidget);
@@ -226,29 +217,27 @@ void main() {
 
     expect(upgrader.isTooSoon(), true);
 
-    expect(find.text(upgrader.messages!.title), findsOneWidget);
+    expect(find.text(upgrader.messages.title), findsOneWidget);
     expect(find.text(upgrader.message()), findsOneWidget);
     expect(find.text('Release Notes:'), findsOneWidget);
     expect(find.text(upgrader.releaseNotes!), findsOneWidget);
-    expect(find.text(upgrader.messages!.prompt), findsOneWidget);
+    expect(find.text(upgrader.messages.prompt), findsOneWidget);
     expect(find.byType(CupertinoDialogAction), findsNWidgets(3));
-    expect(find.text(upgrader.messages!.buttonTitleIgnore), findsOneWidget);
-    expect(find.text(upgrader.messages!.buttonTitleLater), findsOneWidget);
-    expect(find.text(upgrader.messages!.buttonTitleUpdate), findsOneWidget);
+    expect(find.text(upgrader.messages.buttonTitleIgnore), findsOneWidget);
+    expect(find.text(upgrader.messages.buttonTitleLater), findsOneWidget);
+    expect(find.text(upgrader.messages.buttonTitleUpdate), findsOneWidget);
 
-    await tester.tap(find.text(upgrader.messages!.buttonTitleUpdate));
+    await tester.tap(find.text(upgrader.messages.buttonTitleUpdate));
     await tester.pumpAndSettle();
-    expect(find.text(upgrader.messages!.buttonTitleIgnore), findsNothing);
-    expect(find.text(upgrader.messages!.buttonTitleLater), findsNothing);
-    expect(find.text(upgrader.messages!.buttonTitleUpdate), findsNothing);
+    expect(find.text(upgrader.messages.buttonTitleIgnore), findsNothing);
+    expect(find.text(upgrader.messages.buttonTitleLater), findsNothing);
+    expect(find.text(upgrader.messages.buttonTitleUpdate), findsNothing);
     expect(called, true);
     expect(notCalled, true);
   }, skip: false);
   testWidgets('test UpgradeWidget ignore', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader();
-    upgrader.platform = TargetPlatform.iOS;
-    upgrader.client = client;
+    final upgrader = Upgrader(platform: TargetPlatform.iOS, client: client);
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -275,24 +264,22 @@ void main() {
 
     expect(upgrader.isTooSoon(), false);
 
-    await tester.pumpWidget(const _MyWidget());
+    await tester.pumpWidget(_MyWidget(upgrader: upgrader));
 
     // Pump the UI so the upgrader can display its dialog
     await tester.pumpAndSettle();
     await tester.pumpAndSettle();
 
-    await tester.tap(find.text(upgrader.messages!.buttonTitleIgnore));
+    await tester.tap(find.text(upgrader.messages.buttonTitleIgnore));
     await tester.pumpAndSettle();
-    expect(find.text(upgrader.messages!.buttonTitleIgnore), findsNothing);
+    expect(find.text(upgrader.messages.buttonTitleIgnore), findsNothing);
     expect(called, true);
     expect(notCalled, true);
   }, skip: false);
 
   testWidgets('test UpgradeWidget later', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader();
-    upgrader.platform = TargetPlatform.iOS;
-    upgrader.client = client;
+    final upgrader = Upgrader(platform: TargetPlatform.iOS, client: client);
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -319,24 +306,22 @@ void main() {
 
     expect(upgrader.isTooSoon(), false);
 
-    await tester.pumpWidget(const _MyWidget());
+    await tester.pumpWidget(_MyWidget(upgrader: upgrader));
 
     // Pump the UI so the upgrader can display its dialog
     await tester.pumpAndSettle();
     await tester.pumpAndSettle();
 
-    await tester.tap(find.text(upgrader.messages!.buttonTitleLater));
+    await tester.tap(find.text(upgrader.messages.buttonTitleLater));
     await tester.pumpAndSettle();
-    expect(find.text(upgrader.messages!.buttonTitleLater), findsNothing);
+    expect(find.text(upgrader.messages.buttonTitleLater), findsNothing);
     expect(called, true);
     expect(notCalled, true);
   }, skip: false);
 
   testWidgets('test UpgradeWidget pop scope', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader();
-    upgrader.platform = TargetPlatform.iOS;
-    upgrader.client = client;
+    final upgrader = Upgrader(platform: TargetPlatform.iOS, client: client);
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -354,7 +339,7 @@ void main() {
 
     expect(upgrader.isTooSoon(), false);
 
-    await tester.pumpWidget(const _MyWidget());
+    await tester.pumpWidget(_MyWidget(upgrader: upgrader));
 
     // Pump the UI so the upgrader can display its dialog
     await tester.pumpAndSettle();
@@ -363,16 +348,14 @@ void main() {
     // TODO: this test does not pop scope because there is no way to do that.
     // await tester.pageBack();
     // await tester.pumpAndSettle();
-    // expect(find.text(upgrader.messages!.buttonTitleLater), findsNothing);
+    // expect(find.text(upgrader.messages.buttonTitleLater), findsNothing);
     expect(called, false);
   }, skip: false);
 
   testWidgets('test UpgradeWidget Card upgrade', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader();
-    upgrader.client = client;
-    upgrader.platform = TargetPlatform.iOS;
-    upgrader.debugLogging = true;
+    final upgrader = Upgrader(
+        platform: TargetPlatform.iOS, client: client, debugLogging: true);
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -401,27 +384,25 @@ void main() {
 
     expect(upgrader.isTooSoon(), false);
 
-    await tester.pumpWidget(const _MyWidgetCard());
+    await tester.pumpWidget(_MyWidgetCard(upgrader: upgrader));
 
     // Pump the UI so the upgrade card is displayed
     await tester.pumpAndSettle();
 
     expect(find.text('Release Notes:'), findsOneWidget);
     expect(find.text(upgrader.releaseNotes!), findsOneWidget);
-    await tester.tap(find.text(upgrader.messages!.buttonTitleUpdate));
+    await tester.tap(find.text(upgrader.messages.buttonTitleUpdate));
     await tester.pumpAndSettle();
 
     expect(called, true);
     expect(notCalled, true);
-    expect(find.text(upgrader.messages!.buttonTitleUpdate), findsNothing);
+    expect(find.text(upgrader.messages.buttonTitleUpdate), findsNothing);
   }, skip: false);
 
   testWidgets('test UpgradeWidget Card ignore', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader();
-    upgrader.platform = TargetPlatform.iOS;
-    upgrader.client = client;
-    upgrader.debugLogging = true;
+    final upgrader = Upgrader(
+        platform: TargetPlatform.iOS, client: client, debugLogging: true);
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -448,25 +429,23 @@ void main() {
 
     expect(upgrader.isTooSoon(), false);
 
-    await tester.pumpWidget(const _MyWidgetCard());
+    await tester.pumpWidget(_MyWidgetCard(upgrader: upgrader));
 
     // Pump the UI so the upgrade card is displayed
     await tester.pumpAndSettle();
 
-    await tester.tap(find.text(upgrader.messages!.buttonTitleIgnore));
+    await tester.tap(find.text(upgrader.messages.buttonTitleIgnore));
     await tester.pumpAndSettle();
 
     expect(called, true);
     expect(notCalled, true);
-    expect(find.text(upgrader.messages!.buttonTitleIgnore), findsNothing);
+    expect(find.text(upgrader.messages.buttonTitleIgnore), findsNothing);
   }, skip: false);
 
   testWidgets('test UpgradeWidget Card later', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader();
-    upgrader.platform = TargetPlatform.iOS;
-    upgrader.client = client;
-    upgrader.debugLogging = true;
+    final upgrader = Upgrader(
+        platform: TargetPlatform.iOS, client: client, debugLogging: true);
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -493,25 +472,23 @@ void main() {
 
     expect(upgrader.isTooSoon(), false);
 
-    await tester.pumpWidget(const _MyWidgetCard());
+    await tester.pumpWidget(_MyWidgetCard(upgrader: upgrader));
 
     // Pump the UI so the upgrade card is displayed
     await tester.pumpAndSettle(const Duration(milliseconds: 5000));
 
-    await tester.tap(find.text(upgrader.messages!.buttonTitleLater));
+    await tester.tap(find.text(upgrader.messages.buttonTitleLater));
     await tester.pumpAndSettle();
 
     expect(called, true);
     expect(notCalled, true);
-    expect(find.text(upgrader.messages!.buttonTitleLater), findsNothing);
+    expect(find.text(upgrader.messages.buttonTitleLater), findsNothing);
   }, skip: false);
 
   testWidgets('test upgrader minAppVersion', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader();
-    upgrader.platform = TargetPlatform.iOS;
-    upgrader.client = client;
-    upgrader.debugLogging = true;
+    final upgrader = Upgrader(
+        platform: TargetPlatform.iOS, client: client, debugLogging: true);
     upgrader.minAppVersion = '1.0.0';
 
     upgrader.installPackageInfo(
@@ -540,23 +517,21 @@ void main() {
 
     upgrader.minAppVersion = '1.0.0';
 
-    await tester.pumpWidget(const _MyWidgetCard());
+    await tester.pumpWidget(_MyWidgetCard(upgrader: upgrader));
 
     // Pump the UI so the upgrade card is displayed
     await tester.pumpAndSettle(const Duration(milliseconds: 5000));
 
-    expect(find.text(upgrader.messages!.buttonTitleIgnore), findsNothing);
-    expect(find.text(upgrader.messages!.buttonTitleLater), findsNothing);
-    expect(find.text(upgrader.messages!.buttonTitleUpdate), findsOneWidget);
+    expect(find.text(upgrader.messages.buttonTitleIgnore), findsNothing);
+    expect(find.text(upgrader.messages.buttonTitleLater), findsNothing);
+    expect(find.text(upgrader.messages.buttonTitleUpdate), findsOneWidget);
   }, skip: false);
 
   testWidgets('test upgrader minAppVersion description android',
       (WidgetTester tester) async {
     final client = await MockPlayStoreSearchClient.setupMockClient();
-    final upgrader = Upgrader();
-    upgrader.platform = TargetPlatform.android;
-    upgrader.client = client;
-    upgrader.debugLogging = true;
+    final upgrader = Upgrader(
+        platform: TargetPlatform.android, client: client, debugLogging: true);
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -575,10 +550,8 @@ void main() {
     final client = MockITunesSearchClient.setupMockClient(
       description: 'Use this app. [:mav: 4.5.6]',
     );
-    final upgrader = Upgrader();
-    upgrader.platform = TargetPlatform.iOS;
-    upgrader.client = client;
-    upgrader.debugLogging = true;
+    final upgrader = Upgrader(
+        platform: TargetPlatform.iOS, client: client, debugLogging: true);
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -594,11 +567,11 @@ void main() {
 
   testWidgets('test UpgradeWidget unknown app', (WidgetTester tester) async {
     final client = MockITunesSearchClient.setupMockClient();
-    final upgrader = Upgrader();
-    upgrader.platform = TargetPlatform.iOS;
-    upgrader.client = client;
-    upgrader.debugLogging = true;
-    upgrader.countryCode = 'IT';
+    final upgrader = Upgrader(
+        platform: TargetPlatform.iOS,
+        client: client,
+        debugLogging: true,
+        countryCode: 'IT');
 
     upgrader.installPackageInfo(
         packageInfo: PackageInfo(
@@ -625,12 +598,12 @@ void main() {
 
     expect(upgrader.isTooSoon(), false);
 
-    await tester.pumpWidget(const _MyWidgetCard());
+    await tester.pumpWidget(_MyWidgetCard(upgrader: upgrader));
 
     // Pump the UI so the upgrade card is displayed
     await tester.pumpAndSettle();
 
-    final laterButton = find.text(upgrader.messages!.buttonTitleLater);
+    final laterButton = find.text(upgrader.messages.buttonTitleLater);
     expect(laterButton, findsNothing);
 
     expect(called, false);
@@ -641,12 +614,12 @@ void main() {
     test('should use fake Appcast', () async {
       final fakeAppcast = FakeAppcast();
       final client = MockITunesSearchClient.setupMockClient();
-      final upgrader = Upgrader()
-        ..platform = TargetPlatform.iOS
-        ..client = client
-        ..appcastConfig = fakeAppcast.config
-        ..debugLogging = true
-        ..appcast = fakeAppcast
+      final upgrader = Upgrader(
+          platform: TargetPlatform.iOS,
+          client: client,
+          debugLogging: true,
+          appcastConfig: fakeAppcast.config,
+          appcast: fakeAppcast)
         ..installPackageInfo(
           packageInfo: PackageInfo(
             appName: 'Upgrader',
@@ -667,42 +640,47 @@ void main() {
     }, skip: false);
 
     test('durationUntilAlertAgain is 0 days', () async {
-      final upgrader = Upgrader();
-      upgrader.durationUntilAlertAgain = const Duration(seconds: 0);
+      final upgrader =
+          Upgrader(durationUntilAlertAgain: const Duration(seconds: 0));
       expect(upgrader.durationUntilAlertAgain, const Duration(seconds: 0));
 
-      final alert1 =
-          UpgradeAlert(durationToAlertAgain: const Duration(seconds: 0));
-      expect(alert1.durationToAlertAgain, const Duration(seconds: 0));
+      UpgradeAlert(upgrader);
+      expect(upgrader.durationUntilAlertAgain, const Duration(seconds: 0));
 
-      final card1 =
-          UpgradeCard(durationToAlertAgain: const Duration(seconds: 0));
-      expect(card1.durationToAlertAgain, const Duration(seconds: 0));
+      UpgradeCard(upgrader);
+      expect(upgrader.durationUntilAlertAgain, const Duration(seconds: 0));
     }, skip: false);
 
     test('durationUntilAlertAgain card is valid', () async {
-      final card1 = UpgradeCard();
-      expect(card1.durationToAlertAgain, const Duration(days: 3));
-      final card2 = UpgradeCard(durationToAlertAgain: const Duration(days: 10));
-      expect(card2.durationToAlertAgain, const Duration(days: 10));
+      final upgrader =
+          Upgrader(durationUntilAlertAgain: const Duration(days: 3));
+      UpgradeCard(upgrader);
+      expect(upgrader.durationUntilAlertAgain, const Duration(days: 3));
+
+      final upgrader2 =
+          Upgrader(durationUntilAlertAgain: const Duration(days: 10));
+      final _ = UpgradeCard(upgrader2);
+      expect(upgrader2.durationUntilAlertAgain, const Duration(days: 10));
     }, skip: false);
 
     test('durationUntilAlertAgain alert is valid', () async {
-      final alert1 = UpgradeAlert();
-      expect(alert1.durationToAlertAgain, const Duration(days: 3));
-      final alert2 =
-          UpgradeAlert(durationToAlertAgain: const Duration(days: 10));
-      expect(alert2.durationToAlertAgain, const Duration(days: 10));
+      final upgrader =
+          Upgrader(durationUntilAlertAgain: const Duration(days: 3));
+      UpgradeAlert(upgrader);
+      expect(upgrader.durationUntilAlertAgain, const Duration(days: 3));
+
+      final upgrader2 =
+          Upgrader(durationUntilAlertAgain: const Duration(days: 10));
+      UpgradeAlert(upgrader2);
+      expect(upgrader2.durationUntilAlertAgain, const Duration(days: 10));
     }, skip: false);
   });
 
   group('shouldDisplayUpgrade', () {
     test('should respect debugDisplayAlways property', () {
       final client = MockITunesSearchClient.setupMockClient();
-      final upgrader = Upgrader()
-        ..platform = TargetPlatform.iOS
-        ..client = client
-        ..debugLogging = true;
+      final upgrader = Upgrader(
+          platform: TargetPlatform.iOS, client: client, debugLogging: true);
 
       expect(upgrader.shouldDisplayUpgrade(), false);
       upgrader.debugDisplayAlways = true;
@@ -730,10 +708,10 @@ void main() {
     }, skip: false);
 
     test('should return true when version is below minAppVersion', () async {
-      final upgrader = Upgrader()
-        ..client = MockITunesSearchClient.setupMockClient()
-        ..platform = TargetPlatform.iOS
-        ..debugLogging = true
+      final upgrader = Upgrader(
+          debugLogging: true,
+          platform: TargetPlatform.iOS,
+          client: MockITunesSearchClient.setupMockClient())
         ..minAppVersion = '2.0.0'
         ..installPackageInfo(
           packageInfo: PackageInfo(
@@ -752,10 +730,10 @@ void main() {
     }, skip: false);
 
     test('should return true when bestItem has critical update', () async {
-      final upgrader = Upgrader()
-        ..client = MockITunesSearchClient.setupMockClient()
-        ..platform = TargetPlatform.iOS
-        ..debugLogging = true
+      final upgrader = Upgrader(
+          debugLogging: true,
+          platform: TargetPlatform.iOS,
+          client: MockITunesSearchClient.setupMockClient())
         ..installPackageInfo(
           packageInfo: PackageInfo(
             appName: 'Upgrader',
@@ -773,10 +751,10 @@ void main() {
     }, skip: false);
 
     test('packageInfo is empty', () async {
-      final upgrader = Upgrader()
-        ..client = MockITunesSearchClient.setupMockClient()
-        ..platform = TargetPlatform.iOS
-        ..debugLogging = true
+      final upgrader = Upgrader(
+          client: MockITunesSearchClient.setupMockClient(),
+          platform: TargetPlatform.iOS,
+          debugLogging: true)
         ..installPackageInfo(
           packageInfo: PackageInfo(
             appName: '',
@@ -839,11 +817,8 @@ void verifyMessages(UpgraderMessages messages, String code) {
 }
 
 class _MyWidget extends StatelessWidget {
-  final UpgradeDialogStyle dialogStyle;
-  const _MyWidget({
-    Key? key,
-    this.dialogStyle = UpgradeDialogStyle.material,
-  }) : super(key: key);
+  final Upgrader upgrader;
+  const _MyWidget({Key? key, required this.upgrader}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -853,10 +828,7 @@ class _MyWidget extends StatelessWidget {
         appBar: AppBar(
           title: const Text('Upgrader test'),
         ),
-        body: UpgradeAlert(
-            debugLogging: true,
-            dialogStyle: dialogStyle,
-            willDisplayUpgrade: (bool value) {},
+        body: UpgradeAlert(upgrader,
             child: Column(
               children: const <Widget>[Text('Upgrading')],
             )),
@@ -866,9 +838,8 @@ class _MyWidget extends StatelessWidget {
 }
 
 class _MyWidgetCard extends StatelessWidget {
-  const _MyWidgetCard({
-    Key? key,
-  }) : super(key: key);
+  final Upgrader upgrader;
+  const _MyWidgetCard({Key? key, required this.upgrader}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -879,12 +850,7 @@ class _MyWidgetCard extends StatelessWidget {
           title: const Text('Upgrader test'),
         ),
         body: Column(
-          children: <Widget>[
-            UpgradeCard(
-              debugLogging: true,
-              willDisplayUpgrade: (bool value) {},
-            )
-          ],
+          children: <Widget>[UpgradeCard(upgrader)],
         ),
       ),
     );